Import BytesIOSegmentPayload lazily
